Confluent.Kafka.SchemaRegistry -> Confluent.SchemaRegistry
