+++ src/Confluent.SchemaRegistry/Confluent.SchemaRegistry.csproj
