+++ src/Confluent.SchemaRegistry/Rest/RestService.cs
@@ -21,13 +21,13 @@ using System.Net.Http;
 using System.IO;
 using System;
 using System.Threading.Tasks;
-using Confluent.Kafka.SchemaRegistry.Rest.Entities.Requests;
-using Confluent.Kafka.SchemaRegistry.Rest.Entities;
+using Confluent.SchemaRegistry.Rest.Entities.Requests;
+using Confluent.SchemaRegistry.Rest.Entities;
 using Newtonsoft.Json.Linq;
 using Newtonsoft.Json;
 
 
-namespace Confluent.Kafka.SchemaRegistry.Rest
+namespace Confluent.SchemaRegistry.Rest
 {
     /// <remarks>
     ///     It may be useful to expose this publically, but this is not
