+++ src/Confluent.SchemaRegistry/include_docs.xml
