+++ test/Confluent.SchemaRegistry.IntegrationTests/schema.registry.parameters.json
